+++ NEWS.md
@@ -1,5 +1,14 @@
 # NEWS ggboral
 
+Version 0.1.7 2020-06-23
+
+Options for function gg_varpart to format variance component values as
+percentages and append mean component variance to legend label.
+
+Version 0.1.6 2019-03-25
+
+Option to draw caterpillar plots with a single colour.
+
 Version 0.1.5 2018-12-18
 
 Updated to work with boral package version 1.7, in which the number of latent